fix: separate registrations to account for featurelist init
